refactor: move struct nearer to impl
